clean up of files and imports, renamed ComponentFactory to EligibilityPage
