+++ components/EligibilityPage/AccordionFormContainer.tsx
@@ -5,11 +5,11 @@ import type { FormField } from '../../client-state/models/FormField'
 import { WebTranslations } from '../../i18n/web'
 import { FieldType } from '../../utils/api/definitions/fields'
 import { useTranslation } from '../Hooks'
-import { CurrencyField } from './CurrencyField'
-import { NumberField } from './NumberField'
-import { Radio } from './Radio'
-import { FormSelect } from './Select'
-import { TextField } from './TextField'
+import { CurrencyField } from '../Forms/CurrencyField'
+import { NumberField } from '../Forms/NumberField'
+import { Radio } from '../Forms/Radio'
+import { FormSelect } from '../Forms/Select'
+import { TextField } from '../Forms/TextField'
 // import { AccordionForm, Message } from '@dts-stn/decd-design-system'
 import { AccordionForm, Message } from '@solosphere/decd-design-system'
